second commit

+++ routes/groups.py
@@ -112,7 +112,7 @@ def group_dashboard(group_id):
     total_amount = sum(exp.amount for exp in group.expenses)
 
     # Use 0 until Settlement module is ready
-    settlement_count = 0
+    settlement_count = Settlement.query.filter_by(group_id=group.id).count()
 
     # Recent Members
     recent_members = (GroupMember.query.filter_by(group_id=group.id).order_by(GroupMember.joined_at.desc()).limit(5).all())
@@ -217,6 +217,14 @@ def delete_group(group_id):
 
     group = membership.group
 
+    pending_settlements = Settlement.query.filter_by(
+    group_id=group.id,
+    status="Pending").first()
+
+    if pending_settlements:
+        flash("This group has pending settlements. Complete them before deleting the group.","danger")
+        return redirect(url_for("groups.group_dashboard", group_id=group.id))
+
     if request.method == "POST":
 
         db.session.delete(group)
@@ -228,102 +236,6 @@ def delete_group(group_id):
 
     return render_template("groups/delete_group.html", group=group)
 
-
-
-
-
-
-
-
-
-
-
-# ---------------------------------------------------EDIT GROUP--------------------------------------------------------------
-# @groups.route("/<int:group_id>/edit", methods=["GET", "POST"])
-# def edit_group(group_id):
-
-#     if "user_id" not in session:
-#         return redirect(url_for("auth.login"))
-
-#     group = Group.query.filter_by(id=group_id,created_by=session["user_id"]).first_or_404()
-
-#     if request.method == "POST":
-
-#         # Get form data
-#         name = request.form["name"].strip()
-#         description = request.form["description"].strip()
-#         group_type = request.form["group_type"]
-
-#         # Check duplicate group name
-#         existing = Group.query.filter(
-#             Group.name == name,
-#             Group.created_by == session["user_id"],
-#             Group.id != group.id
-#         ).first()
-
-#         if existing:
-#             flash("Group name already exists.", "warning")
-#             return redirect(
-#                 url_for("groups.edit_group", group_id=group.id)
-#             )
-
-#         # Update group
-#         group.name = name
-#         group.description = description
-#         group.group_type = group_type
-
-#         # Trip Mode
-#         if "trip_mode" in request.form:
-
-#             start_date = request.form.get("trip_start_date")
-#             end_date = request.form.get("trip_end_date")
-
-#             if not start_date or not end_date:
-#                 flash("Trip dates are required when Trip Mode is enabled.","danger")
-#                 return redirect(url_for("groups.edit_group", group_id=group.id))
-
-#             group.trip_mode = True
-#             group.trip_start_date = datetime.strptime(start_date,"%Y-%m-%d").date()
-#             group.trip_end_date = datetime.strptime(end_date,"%Y-%m-%d").date()
-
-#             if group.trip_end_date < group.trip_start_date:
-#                 flash("Trip end date cannot be before start date.","danger")
-#                 return redirect(url_for("groups.edit_group", group_id=group.id))
-#         else:
-#             group.trip_mode = False
-#             group.trip_start_date = None
-#             group.trip_end_date = None
-            
-#         db.session.commit()
-#         flash("Group updated successfully.", "success")
-
-#         return redirect(url_for("groups.group_dashboard", group_id=group.id))
-#     return render_template("groups/edit_group.html",group=group)
-
-
-# ----------------------------------------------------DELETE GROUP----------------------------------------------------
-
-# @groups.route("/<int:group_id>/delete", methods=["GET", "POST"])
-# def delete_group(group_id):
-
-#     if "user_id" not in session:
-#         return redirect(url_for("auth.login"))
-
-#     group = Group.query.filter_by(
-#         id=group_id,
-#         created_by=session["user_id"]
-#     ).first_or_404()
-
-#     if request.method == "POST":
-
-#         db.session.delete(group)
-#         db.session.commit()
-
-#         flash("Group deleted successfully.", "success")
-
-#         return redirect(url_for("groups.groups_home"))
-
-#     return render_template("groups/delete_group.html",group=group)
 
 # ----------------------------------------------------GROUP MEMBERS----------------------------------------------------
